add more tests

+++ src/main/kotlin/dev/drf/midas/glove/core/motion/Velocity.kt
@@ -1,3 +1,3 @@
 package dev.drf.midas.glove.core.motion
 
-data class Velocity(val power: Long)
+data class Velocity(val power: Double)
